Refactor: Move gmailConfig to companion object
This commit moves the `gmailConfig` variable from the inner `UserMailServiceTests` class to the companion object of the outer `Tests` class. This makes it accessible to other inner classes within `Tests`, such as `ContextTests`.

This change improves code organization and reusability by making the `gmailConfig` available to multiple test classes that require it.