feat: Remove mail-based activation key validation and update Utils
This commit removes the mail-based activation key validation from the `Tests.kt` file. The previous logic that retrieved and validated an activation key from an email is now commented out. The change was made because the reliance on Gmail was problematic as not every test sent an email, leading to unreliable test behavior. The commit also introduces a minor improvement to `Utils.kt` by making the code a little more concise and adding an empty line.

- Removed the conditional block for email validation in `Tests.kt`, the code is commented out.
- Update `Utils.kt`, add empty line.
- Update `Utils.kt`, code more concise.